It is my first commit for this repository
I wrote this web application project together with my team at our university. We used HTML, CSS, SASS. This site already exists, we took up writing this web page only to strengthen our teamwork skills, as well as to consolidate knowledge in the field of frontend developer. We did not copy anyone's code, we took only logos and photos, our team is not responsible for this project. And we have nothing to do with the representatives of this site and organization.

Our project consists of several pages, according to the standards of the current design structure by 2022. The page is adapted to the screen size of three devices, such as standard laptops and iPad Air and Samsung Galaxy A71. We used HTMl and SASS for page layout, then we converted SASS to CSS. And we added a little animation using regular CSS.